model.ts added
