+++ frontend/src/Components/TodoItem.tsx
@@ -1,14 +1,12 @@
+import {Todo} from "./model";
+import './TodoItem.css';
 
-export interface Todo {
-    id?: string;
-    task: string;
-    description?: string;
-    status: string;
-}
+
 
 function TodoItem(props: Todo) {
     return(
         <div>
+
         </div>
     )
 }
