+++ frontend/src/Components/model.ts
@@ -1,0 +1,6 @@
+export interface Todo {
+    id?: string;
+    task: string;
+    description?: string;
+    status: string;
+}
